# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY Orange Money API",
        "description": "API pour l'application Orange Money - OMPAY",
        "contact": {
            "email": "support@orangemoney.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.ompaysn.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/auth/creercompte": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Initier l'inscription d'un nouvel utilisateur",
                "description": "Crée un compte utilisateur en envoyant un OTP pour vérification",
                "operationId": "initierInscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prenom",
                                    "nom",
                                    "numeroTelephone"
                                ],
                                "properties": {
                                    "prenom": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "nom": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "numeroTelephone": {
                                        "description": "Numéro de téléphone au format +221XXXXXXXXX",
                                        "type": "string",
                                        "example": "+221771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP envoyé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "idUtilisateur": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "numeroTelephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "VALIDATION_ERROR"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Données invalides"
                                                },
                                                "details": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/verification-otp": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier le code OTP pour connexion",
                "description": "Valide le code OTP pour permettre la connexion d'un utilisateur existant",
                "operationId": "verificationOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numeroTelephone",
                                    "codeOTP"
                                ],
                                "properties": {
                                    "numeroTelephone": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "codeOTP": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP validé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP validé"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "refreshToken": {
                                                    "type": "string",
                                                    "example": "refresh_token_here"
                                                },
                                                "expiresIn": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code OTP invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "OTP_INVALID"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code OTP invalide ou expiré"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/deconnexion": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Invalide le token d'accès actuel",
                "operationId": "deconnexion",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}